Add files via upload
This is the training process. We use the data.yaml and model.yaml to require our model and data input format. Notice that I adjust the weights of weed and crop, since I find that each time I run the model, the accuracy of weed less than crop obviously. To solve the problem, we increase the weights of crop so that it can converge faster.

Besides, I make an announcement about model train format: the optimizer and learn rate o and f are given after some attempts.