Add files via upload
This image is the results of my training process. As we can see, the loss of box and classification is not so ideal. Even if I have already used some Methods of data augmentation to expand it's dataset. But this image set is still too small, which only contains 1400 images. 

We can find a working regulation, the efficiency of deep learning comes from the large dataset. It reflects the value of deep learning network per data is till too low. If you from this efficiency perspective, we can conclude that deep learning a low efficiency data mining technology compared with other machine learning algorithm.